% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_coord}
\alias{convert_coord}
\title{Convert coordinate of from DDMM.mmmm to decimal degrees}
\usage{
convert_coord(x)
}
\arguments{
\item{x}{Longitude or latitude numeric vector}
}
\value{
A vector
}
\description{
Convert coordinate of from DDMM.mmmm to decimal degrees
}
