% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_coord_DDMM.mmmmm}
\alias{convert_coord_DDMM.mmmmm}
\title{Convert coordinate of from DDMM.mmmm to decimal degrees}
\usage{
convert_coord_DDMM.mmmmm(x)
}
\arguments{
\item{x}{Longitude or latitude numeric vector}
}
\value{
A vector
}
\description{
Convert coordinate of from DDMM.mmmm to decimal degrees
}
