% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{hr_connection}
\alias{hr_connection}
\title{Connetion to hafriti}
\usage{
hr_connection(password)
}
\arguments{
\item{password}{Your password}
}
\value{
A SQL connection
}
\description{
Connetion to hafriti
}
