% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hr_fields}
\alias{hr_fields}
\title{Informations on fields}
\format{
\subsection{\code{hr_fields}}{

A data frame with 104 rows and 3 columns:
\describe{
\item{type}{type code}
\item{label}{label name}
\item{name}{variable name}
\item{class}{variable class}
\item{details}{information on values}
}
}
}
\usage{
hr_fields
}
\description{
Informations on fields
}
\keyword{datasets}
