% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{hr_message_parse}
\alias{hr_message_parse}
\title{Parse messages}
\usage{
hr_message_parse(m, TYPE = "ATW")
}
\arguments{
\item{m}{A dataframe containing message field, now limited to "m"}

\item{TYPE}{A character, containing messge type}
}
\value{
A tibble containg parsed messge
}
\description{
Parse messages
}
