% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor.R
\name{hr_sensor}
\alias{hr_sensor}
\title{Hafriti sensor data}
\usage{
hr_sensor(con, skip = "arni")
}
\arguments{
\item{con}{a database connection}

\item{skip}{vessel name (default: 'arni')}
}
\value{
A SQL connection
}
\description{
Arni and Bjarni are in separate table
}
