% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hr_talkers}
\alias{hr_talkers}
\title{Informations on talkers}
\format{
\subsection{\code{hr_talkers}}{

A data frame with 50 rows and 3 columns:
\describe{
\item{talker}{talker code}
\item{talker_desc}{talker description}
\item{source}{data source}
}
}
}
\usage{
hr_talkers
}
\description{
Informations on talkers
}
\keyword{datasets}
