% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hr_types}
\alias{hr_types}
\title{Informations on types}
\format{
\subsection{\code{hr_types}}{

A data frame with 104 rows and 3 columns:
\describe{
\item{type}{type code}
\item{type_desc}{type description}
\item{source}{data source}
}
}
}
\usage{
hr_types
}
\description{
Informations on types
}
\keyword{datasets}
