% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postgres_test_data}
\alias{postgres_test_data}
\title{postres test data}
\format{
\subsection{\code{postgres_test_data}}{

A data frame with 104 rows and 3 columns:
\describe{
\item{id}{type code}
\item{channel_id}{label name}
\item{date-time}{variable name}
\item{message_type}{variable class}
\item{nmea_message}{information on values}
\item{json_message}{xxx}
\item{talker}{derived from message_type and nmea_message}
\item{type}{derived from message_type and nmea_message}
}
}
}
\usage{
postgres_test_data
}
\description{
Dataset that contains dumps from Arni and Bjarni, 100 records for each
message type
}
\keyword{datasets}
