% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet.R
\name{read_speed}
\alias{read_speed}
\title{Title}
\usage{
read_speed(
  vid = 2350,
  t1 = "2024-02-27",
  t2 = "2024-03-21",
  mean_by = "sec",
  path = "/home/haf/einarhj/stasi/hafriti/data/type=VTG"
)
}
\arguments{
\item{vid}{x}

\item{t1}{x}

\item{t2}{x}

\item{mean_by}{x}

\item{path}{x}
}
\value{
A tibble
}
\description{
Title
}
